requirements.txt
streamlit
streamlit-image-select
streamlit_image_coordinates
pandas
Pillow
